added bets

+++ PSA/Shared/Bet.cs
@@ -10,7 +10,7 @@ namespace PSA.Shared
     {
         public int Id { get; set; }
         public double Amount { get; set; }
-		public int Coefficient { get; set; }
+		public double Coefficient { get; set; }
 		public int fk_robot_id { get; set; }
 		public int fk_fight_id { get; set; }
 		public int fk_user_id { get; set; }
